move NodeInfo to compute module
